requirements.txt
Django
djangorestframework
django-cors-headers
dj-database-url
python-dotenv
psycopg[binary]
rav
python-decouple
gunicorn